move checkbox label to child elements instead of label prop and pass props to children
