+++ lib/Styled/Checkbox/Checkbox.tsx
@@ -24,10 +24,11 @@ const Checkbox = memo(
       defaultChecked = false,
       isIndeterminate = false,
       onChange: onChangeProps,
-      label,
       title,
       name,
       value,
+      children,
+      textProps,
       ...rest
     } = props;
 
@@ -49,6 +50,21 @@ const Checkbox = memo(
     const isChecked =
       isCheckedProp === undefined ? isCheckedState : isCheckedProp;
     const id = useUID();
+
+    // Add props to children
+    const childrenWithProps = React.Children.map(children, child => {
+      // Checking isValidElement is the safe way and avoids a typescript
+      // error too.
+      if (React.isValidElement(child)) {
+        return React.cloneElement(child, {
+          isDisabled,
+          isChecked,
+          style: { fontSize: "inherit" }
+        });
+      }
+      return child;
+    });
+
     return (
       <TextSpan
         as={"label"}
@@ -73,7 +89,7 @@ const Checkbox = memo(
             cursor: not-allowed;
           `}
         `}
-        {...rest}
+        {...textProps}
       >
         <HiddenCheckbox
           disabled={isDisabled}
@@ -90,14 +106,7 @@ const Checkbox = memo(
           isDisabled={isDisabled}
         />
         <SpacingSpan right={1} />
-        <TextSpan
-          isDisabled={isDisabled}
-          css={`
-            font-size: inherit;
-          `}
-        >
-          {label}
-        </TextSpan>
+        {childrenWithProps}
       </TextSpan>
     );
   })
